+++ fe/src/pages/Template.tsx
@@ -5,7 +5,7 @@ import Button from 'components/common/Button/Button';
 import ButtonLink from 'components/common/Button/ButtonLink';
 import DropdownBtn from 'components/common/DropdownBtn';
 import Icon from 'components/common/Icon';
-import Input from 'components/common/Input';
+import Input from 'components/common/Input/Input';
 import Label from 'components/common/Label';
 import LinearGraph from 'components/common/LinearGraph';
 import ListContainer from 'components/common/ListContainer';
